requirements.txt
asgiref==3.3.1
Django==3.1.4
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.14
Pillow==8.0.1
psycopg2-binary==2.8.6
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1